whatsapp-clone-backend/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.22.0
motor==3.7.1
pydantic==1.10.13
python-multipart==0.0.6
python-dotenv==1.0.0
